requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu # for torch for cpu only devices (no gpu)

SpeechRecognition
PyAudio
google-generativeai
gpt4all
gtts
pyttsx3
psutil
vosk
torch
